third_party/blink/web_tests/http/tests/websocket/handshake-fail-by-prepended-null-expected.txt

Connection should fail immediately, rather than succeeding or staying in limbo until timeout, if a null byte is received before the handshake.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS timedOut is false
PASS connected is false
PASS wsOrigin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
